+++ src/plugins/intel_cpu/src/nodes/experimental_detectron_detection_output.cpp
@@ -275,7 +275,7 @@ ExperimentalDetectronDetectionOutput::ExperimentalDetectronDetectionOutput(const
     max_delta_log_wh_ = attributes.max_delta_log_wh;
     classes_num_ = attributes.num_classes;
     max_detections_per_class_ = attributes.post_nms_count;
-    max_detections_per_image_ = attributes.max_detections_per_image;
+    max_detections_per_image_ = static_cast<int>(attributes.max_detections_per_image);
     class_agnostic_box_regression_ = attributes.class_agnostic_box_regression;
     deltas_weights_ = attributes.deltas_weights;
 }
@@ -299,7 +299,7 @@ void ExperimentalDetectronDetectionOutput::initSupportedPrimitiveDescriptors() {
 }
 
 void ExperimentalDetectronDetectionOutput::execute([[maybe_unused]] const dnnl::stream& strm) {
-    const int rois_num = getParentEdgeAt(INPUT_ROIS)->getMemory().getStaticDims()[0];
+    const int rois_num = static_cast<int>(getParentEdgeAt(INPUT_ROIS)->getMemory().getStaticDims()[0]);
     assert(classes_num_ == static_cast<int>(getParentEdgeAt(INPUT_SCORES)->getMemory().getStaticDims()[1]));
     assert(4 * classes_num_ == static_cast<int>(getParentEdgeAt(INPUT_DELTAS)->getMemory().getStaticDims()[1]));
