+++ src/plugins/intel_cpu/src/nodes/gather_elements.cpp
@@ -69,16 +69,16 @@ void GatherElements::prepareParams() {
     const auto& dstDims = getChildEdgeAt(0)->getMemory().getStaticDims();
     strideAxDst_ = 1;
     for (size_t i = dstDims.size() - 1; i > axis_; i--) {
-        strideAxDst_ *= dstDims[i];
+        strideAxDst_ *= static_cast<int>(dstDims[i]);
     }
-    dstAxDim_ = dstDims[axis_];
-    dataAxDim_ = dataDims[axis_];
+    dstAxDim_ = static_cast<int>(dstDims[axis_]);
+    dataAxDim_ = static_cast<int>(dataDims[axis_]);
     if (axis_ > 0) {
         strideAx1Diff_ = 1;
         for (size_t i = dataDims.size() - 1; i >= axis_; i--) {
-            strideAx1Diff_ *= dataDims[i];
+            strideAx1Diff_ *= static_cast<int>(dataDims[i]);
         }
-        strideAx1Diff_ -= strideAxDst_ * dstDims[axis_];
+        strideAx1Diff_ -= strideAxDst_ * static_cast<int>(dstDims[axis_]);
     }
 }
 
@@ -125,7 +125,7 @@ void GatherElements::directExecution() {
     const auto* indices = getSrcDataAtPortAs<const int>(indicesIndex_);
     auto* dstData = getDstDataAtPortAs<dataType>(0);
 
-    const int outSize = getChildEdgeAt(0)->getMemory().getShape().getElementsCount();
+    const int outSize = static_cast<int>(getChildEdgeAt(0)->getMemory().getShape().getElementsCount());
     auto threadBody = [&](const int ithr, const int nthr) {
         int start(0LU);
         int end(0LU);
